Crate reth_provider

Collection of traits and trait implementations for common database operations.

Feature Flags

test-utils: Export utilities for testing

Re-exports

pub use providers::DatabaseProvider;

pub use providers::DatabaseProviderRO;

pub use providers::DatabaseProviderRW;

pub use providers::HistoricalStateProvider;

pub use providers::HistoricalStateProviderRef;

pub use providers::LatestStateProvider;

pub use providers::LatestStateProviderRef;

pub use providers::ProviderFactory;

pub use chain::Chain;

pub use chain::DisplayBlocksChain;

pub use bundle_state::BundleStateWithReceipts;

pub use bundle_state::StateChanges;

pub use bundle_state::StateReverts;

Modules

bundle_state

Bundle state module. This module contains all the logic related to bundle state.

chain

Contains Chain, a chain of blocks and their final state.

providers

Provider trait implementations.

test_utils

Common test helpers for mocking the Provider.

Structs

BlockExecutorStats

Block execution statistics. Contains duration of each step of block execution.

CanonStateNotificationStream

A Stream of CanonStateNotification.

HeaderSyncGap

Represents a gap to sync: from local_head to target

Enums

BlockSource

A helper enum that represents the origin of the requested block.

CanonStateNotification

Chain action that is triggered when a new block is imported or old block is reverted. and will return all crate::BundleStateWithReceipts and reth_primitives::SealedBlockWithSenders of both reverted and committed blocks.

HeaderSyncMode

The header sync mode.

OriginalValuesKnown

Option for BundleState when converting it to the plain state.

ProviderError

Re-export provider error. Bundled errors variants thrown by various providers.

TransactionVariant

Enum to control transaction hash inclusion.

Traits

AccountExtReader

Account reader

AccountReader

Account reader

BlockExecutionWriter

BlockExecution Writer

BlockExecutor

An executor capable of executing a block.

BlockHashReader

Client trait for fetching block hashes by number.

BlockIdReader

Client trait for transforming BlockId into block numbers or hashes.

BlockNumReader

Client trait for getting important block numbers (such as the latest block number), converting block hashes to numbers, and fetching a block hash from its block number.

BlockReader

Api trait for fetching Block related data.

BlockReaderIdExt

Trait extension for BlockReader, for types that implement BlockId conversion.

BlockWriter

Block Writer

BlockchainTreePendingStateProvider

Blockchain trait provider that gives access to the blockchain state that is not yet committed (pending).

BundleStateDataProvider

Post state data needs for execution on it. This trait is used to create a state provider over pending state.

CanonChainTracker

A type that can track updates related to fork choice updates.

CanonStateSubscriptions

A type that allows to register chain related event subscriptions.

ChainSpecProvider

A trait for reading the current chainspec.

ChangeSetReader

AccountChange reader

EvmEnvProvider

A provider type that knows chain specific information required to configure an Env.

ExecutorFactory

Executor factory that would create the EVM with particular state provider.

HashingWriter

Hashing Writer

HeaderProvider

Client trait for fetching Header related data.

HeaderSyncGapProvider

Client trait for determining the current headers sync gap.

HistoryWriter

History Writer

PrunableBlockExecutor

A BlockExecutor capable of in-memory pruning of the data that will be written to the database.

PruneCheckpointReader

The trait for fetching prune checkpoint related data.

PruneCheckpointWriter

The trait for updating prune checkpoint related data.

ReceiptProvider

Client trait for fetching Receipt data .

ReceiptProviderIdExt

Trait extension for ReceiptProvider, for types that implement BlockId conversion.

StageCheckpointReader

The trait for fetching stage checkpoint related data.

StageCheckpointWriter

The trait for updating stage checkpoint related data.

StateProvider

An abstraction for a type that provides state data.

StateProviderFactory

Light wrapper that returns StateProvider implementations that correspond to the given BlockNumber, the latest state, or the pending state.

StateRootProvider

A type that can compute the state root of a given post state.

StorageReader

Storage reader

TransactionsProvider

Client trait for fetching TransactionSigned related data.

TransactionsProviderExt

Client trait for fetching additional TransactionSigned related data.

WithdrawalsProvider

Client trait for fetching Withdrawal related data.

Type Aliases

CanonStateNotificationSender

Type alias for a sender that sends CanonStateNotification

CanonStateNotifications

Type alias for a receiver that receives CanonStateNotification

StateProviderBox

Type alias of boxed StateProvider.