Struct reth_provider::providers::BlockchainProvider

pub struct BlockchainProvider<DB, Tree> { /* private fields */ }

The main type for interacting with the blockchain.

This type serves as the main entry point for interacting with the blockchain and provides data from database storage and from the blockchain tree (pending state etc.) It is a simple wrapper type that holds an instance of the database and the blockchain tree.

Implementations

impl<DB, Tree> BlockchainProvider<DB, Tree>

pub fn with_latest( database: ProviderFactory<DB>, tree: Tree, latest: SealedHeader ) -> Self

Create new provider instance that wraps the database and the blockchain tree, using the provided latest header to initialize the chain info tracker.

impl<DB, Tree> BlockchainProvider<DB, Tree>
where DB: Database,

pub fn new(database: ProviderFactory<DB>, tree: Tree) -> ProviderResult<Self>

Create a new provider using only the database and the tree, fetching the latest header from the database to initialize the provider.