Struct reth_provider::providers::DatabaseProvider

pub struct DatabaseProvider<TX> { /* private fields */ }

A provider struct that fetchs data from the database. Wrapper around DbTx and DbTxMut. Example: HeaderProvider BlockHashReader

Implementations

impl<TX: DbTxMut> DatabaseProvider<TX>

pub fn new_rw(tx: TX, chain_spec: Arc<ChainSpec>) -> Self

Creates a provider with an inner read-write transaction.

impl<TX: DbTx> DatabaseProvider<TX>

pub fn new(tx: TX, chain_spec: Arc<ChainSpec>) -> Self

Creates a provider with an inner read-only transaction.

pub fn with_snapshot_provider( self, snapshot_provider: Arc<SnapshotProvider> ) -> Self

Creates a new Self with access to a SnapshotProvider.

pub fn into_tx(self) -> TX

Consume DbTx or DbTxMut.

pub fn tx_mut(&mut self) -> &mut TX

Pass DbTx or DbTxMut mutable reference.

pub fn tx_ref(&self) -> &TX

Pass DbTx or DbTxMut immutable reference.

pub fn table<T: Table>(&self) -> Result<Vec<KeyValue<T>>, DatabaseError>
where T::Key: Default + Ord,

Return full table as Vec

impl<TX: DbTxMut + DbTx> DatabaseProvider<TX>

pub fn commit(self) -> ProviderResult<bool>

Commit database transaction.

pub fn get_or_take<T: Table, const TAKE: bool>( &self, range: impl RangeBounds<T::Key> ) -> Result<Vec<KeyValue<T>>, DatabaseError>

Return list of entries from table

If TAKE is true, opened cursor would be write and it would delete all values from db.

pub fn unwind_table_by_num<T>(&self, num: u64) -> Result<usize, DatabaseError>
where T: Table<Key = u64>,

Unwind table by some number key. Returns number of rows unwound.

Note: Key is not inclusive and specified key would stay in db.

pub fn unwind_table_by_walker<T1, T2>( &self, start_at: T1::Key ) -> Result<(), DatabaseError>
where T1: Table, T2: Table<Key = T1::Value>,

Unwind a table forward by a Walker on another table

pub fn prune_table_with_iterator<T: Table>( &self, keys: impl IntoIterator<Item = T::Key>, limit: usize, delete_callback: impl FnMut(TableRow<T>) ) -> Result<(usize, bool), DatabaseError>

Prune the table for the specified pre-sorted key iterator.

Returns number of rows pruned.

pub fn prune_table_with_range<T: Table>( &self, keys: impl RangeBounds<T::Key> + Clone + Debug, limit: usize, skip_filter: impl FnMut(&TableRow<T>) -> bool, delete_callback: impl FnMut(TableRow<T>) ) -> Result<(usize, bool), DatabaseError>

Prune the table for the specified key range.

Returns number of rows pruned.