Trait reth_provider::BlockReaderIdExt

pub trait BlockReaderIdExt: BlockReader + BlockIdReader + ReceiptProviderIdExt {

    // Required methods
    fn block_by_id(&self, id: BlockId) -> ProviderResult<Option<Block>>;
    fn sealed_header_by_id(
        &self,
        id: BlockId
    ) -> ProviderResult<Option<SealedHeader>>;
    fn header_by_id(&self, id: BlockId) -> ProviderResult<Option<Header>>;
    fn ommers_by_id(&self, id: BlockId) -> ProviderResult<Option<Vec<Header>>>;

    // Provided methods
    fn block_by_number_or_tag(
        &self,
        id: BlockNumberOrTag
    ) -> ProviderResult<Option<Block>> { ... }
    fn pending_header(&self) -> ProviderResult<Option<SealedHeader>> { ... }
    fn latest_header(&self) -> ProviderResult<Option<SealedHeader>> { ... }
    fn safe_header(&self) -> ProviderResult<Option<SealedHeader>> { ... }
    fn finalized_header(&self) -> ProviderResult<Option<SealedHeader>> { ... }
    fn block_with_senders_by_id(
        &self,
        id: BlockId,
        transaction_kind: TransactionVariant
    ) -> ProviderResult<Option<BlockWithSenders>> { ... }
    fn header_by_number_or_tag(
        &self,
        id: BlockNumberOrTag
    ) -> ProviderResult<Option<Header>> { ... }
    fn sealed_header_by_number_or_tag(
        &self,
        id: BlockNumberOrTag
    ) -> ProviderResult<Option<SealedHeader>> { ... }
    fn ommers_by_number_or_tag(
        &self,
        id: BlockNumberOrTag
    ) -> ProviderResult<Option<Vec<Header>>> { ... }

}

Trait extension for BlockReader, for types that implement BlockId conversion.

The BlockReader trait should be implemented on types that can retrieve a block from either a block number or hash. However, it might be desirable to fetch a block from a BlockId type, which can be a number, hash, or tag such as BlockNumberOrTag::Safe.

Resolving tags requires keeping track of block hashes or block numbers associated with the tag, so this trait can only be implemented for types that implement BlockIdReader. The BlockIdReader methods should be used to resolve BlockIds to block numbers or hashes, and retrieving the block should be done using the type’s BlockReader methods.

Required Methods

fn block_by_id(&self, id: BlockId) -> ProviderResult<Option<Block>>

Returns the block with the matching BlockId from the database.

Returns None if block is not found.

fn sealed_header_by_id( &self, id: BlockId ) -> ProviderResult<Option<SealedHeader>>

Returns the sealed header with the matching BlockId from the database.

Returns None if header is not found.

fn header_by_id(&self, id: BlockId) -> ProviderResult<Option<Header>>

Returns the header with the matching BlockId from the database.

Returns None if header is not found.

fn ommers_by_id(&self, id: BlockId) -> ProviderResult<Option<Vec<Header>>>

Returns the ommers with the matching BlockId from the database.

Returns None if block is not found.

Provided Methods

fn block_by_number_or_tag( &self, id: BlockNumberOrTag ) -> ProviderResult<Option<Block>>

Returns the block with matching tag from the database

Returns None if block is not found.

fn pending_header(&self) -> ProviderResult<Option<SealedHeader>>

Returns the pending block header if available

Note: This returns a SealedHeader because it’s expected that this is sealed by the provider and the caller does not know the hash.

fn latest_header(&self) -> ProviderResult<Option<SealedHeader>>

Returns the latest block header if available

Note: This returns a SealedHeader because it’s expected that this is sealed by the provider and the caller does not know the hash.

fn safe_header(&self) -> ProviderResult<Option<SealedHeader>>

Returns the safe block header if available

Note: This returns a SealedHeader because it’s expected that this is sealed by the provider and the caller does not know the hash.

fn finalized_header(&self) -> ProviderResult<Option<SealedHeader>>

Returns the finalized block header if available

Note: This returns a SealedHeader because it’s expected that this is sealed by the provider and the caller does not know the hash.

fn block_with_senders_by_id( &self, id: BlockId, transaction_kind: TransactionVariant ) -> ProviderResult<Option<BlockWithSenders>>

Returns the block with senders with matching BlockId.

Returns the block’s transactions in the requested variant.

Returns None if block is not found.

fn header_by_number_or_tag( &self, id: BlockNumberOrTag ) -> ProviderResult<Option<Header>>

Returns the header with matching tag from the database

Returns None if header is not found.

fn sealed_header_by_number_or_tag( &self, id: BlockNumberOrTag ) -> ProviderResult<Option<SealedHeader>>

Returns the header with matching tag from the database

Returns None if header is not found.

fn ommers_by_number_or_tag( &self, id: BlockNumberOrTag ) -> ProviderResult<Option<Vec<Header>>>

Returns the ommers with the matching tag from the database.

Object Safety

This trait is not object safe.