Trait reth_provider::BlockWriter

pub trait BlockWriter: Send + Sync {
    // Required methods
    fn insert_block(
        &self,
        block: SealedBlockWithSenders,
        prune_modes: Option<&PruneModes>
    ) -> ProviderResult<StoredBlockBodyIndices>;
    fn append_blocks_with_state(
        &self,
        blocks: Vec<SealedBlockWithSenders>,
        state: BundleStateWithReceipts,
        hashed_state: HashedPostState,
        trie_updates: TrieUpdates,
        prune_modes: Option<&PruneModes>
    ) -> ProviderResult<()>;
}

Block Writer

Required Methods

fn insert_block( &self, block: SealedBlockWithSenders, prune_modes: Option<&PruneModes> ) -> ProviderResult<StoredBlockBodyIndices>

Insert full block and make it canonical. Parent tx num and transition id is taken from parent block in database.

Return StoredBlockBodyIndices that contains indices of the first and last transactions and transition in the block.

fn append_blocks_with_state( &self, blocks: Vec<SealedBlockWithSenders>, state: BundleStateWithReceipts, hashed_state: HashedPostState, trie_updates: TrieUpdates, prune_modes: Option<&PruneModes> ) -> ProviderResult<()>

Appends a batch of sealed blocks to the blockchain, including sender information, and updates the post-state.

Inserts the blocks into the database and updates the state with provided BundleState.

Parameters

blocks: Vector of SealedBlockWithSenders instances to append.

state: Post-state information to update after appending.

prune_modes: Optional pruning configuration.

Returns

Returns Ok(()) on success, or an error if any operation fails.