Trait reth_provider::StateRootProvider

pub trait StateRootProvider: Send + Sync {
    // Required methods
    fn state_root(
        &self,
        bundle_state: &BundleStateWithReceipts
    ) -> ProviderResult<B256>;
    fn state_root_with_updates(
        &self,
        bundle_state: &BundleStateWithReceipts
    ) -> ProviderResult<(B256, TrieUpdates)>;
}

A type that can compute the state root of a given post state.

Required Methods

fn state_root( &self, bundle_state: &BundleStateWithReceipts ) -> ProviderResult<B256>

Returns the state root of the BundleState on top of the current state.

NOTE: It is recommended to provide a different implementation from state_root_with_updates since it affects the memory usage during state root computation.

fn state_root_with_updates( &self, bundle_state: &BundleStateWithReceipts ) -> ProviderResult<(B256, TrieUpdates)>

Returns the state root of the BundleState on top of the current state with trie updates to be committed to the database.