Trait reth_provider::AccountExtReader

pub trait AccountExtReader: Send + Sync {
    // Required methods
    fn changed_accounts_with_range(
        &self,
        _range: impl RangeBounds<BlockNumber>
    ) -> ProviderResult<BTreeSet<Address>>;
    fn basic_accounts(
        &self,
        _iter: impl IntoIterator<Item = Address>
    ) -> ProviderResult<Vec<(Address, Option<Account>)>>;
    fn changed_accounts_and_blocks_with_range(
        &self,
        range: RangeInclusive<BlockNumber>
    ) -> ProviderResult<BTreeMap<Address, Vec<BlockNumber>>>;
}

Account reader

Required Methods

fn changed_accounts_with_range( &self, _range: impl RangeBounds<BlockNumber> ) -> ProviderResult<BTreeSet<Address>>

Iterate over account changesets and return all account address that were changed.

fn basic_accounts( &self, _iter: impl IntoIterator<Item = Address> ) -> ProviderResult<Vec<(Address, Option<Account>)>>

Get basic account information for multiple accounts. A more efficient version than calling AccountReader::basic_account repeatedly.

Returns None if the account doesn’t exist.

fn changed_accounts_and_blocks_with_range( &self, range: RangeInclusive<BlockNumber> ) -> ProviderResult<BTreeMap<Address, Vec<BlockNumber>>>

Iterate over account changesets and return all account addresses that were changed alongside each specific set of blocks.

NOTE: Get inclusive range of blocks.

Object Safety

This trait is not object safe.