Trait reth_provider::EvmEnvProvider

pub trait EvmEnvProvider: Send + Sync {
    // Required methods
    fn fill_env_at<EvmConfig>(
        &self,
        cfg: &mut CfgEnv,
        block_env: &mut BlockEnv,
        at: BlockHashOrNumber,
        evm_config: EvmConfig
    ) -> ProviderResult<()>
       where EvmConfig: ConfigureEvmEnv;
    fn fill_env_with_header<EvmConfig>(
        &self,
        cfg: &mut CfgEnv,
        block_env: &mut BlockEnv,
        header: &Header,
        evm_config: EvmConfig
    ) -> ProviderResult<()>
       where EvmConfig: ConfigureEvmEnv;
    fn fill_block_env_at(
        &self,
        block_env: &mut BlockEnv,
        at: BlockHashOrNumber
    ) -> ProviderResult<()>;
    fn fill_block_env_with_header(
        &self,
        block_env: &mut BlockEnv,
        header: &Header
    ) -> ProviderResult<()>;
    fn fill_cfg_env_at<EvmConfig>(
        &self,
        cfg: &mut CfgEnv,
        at: BlockHashOrNumber,
        evm_config: EvmConfig
    ) -> ProviderResult<()>
       where EvmConfig: ConfigureEvmEnv;
    fn fill_cfg_env_with_header<EvmConfig>(
        &self,
        cfg: &mut CfgEnv,
        header: &Header,
        evm_config: EvmConfig
    ) -> ProviderResult<()>
       where EvmConfig: ConfigureEvmEnv;

    // Provided method
    fn env_with_header<EvmConfig>(
        &self,
        header: &Header,
        evm_config: EvmConfig
    ) -> ProviderResult<(CfgEnv, BlockEnv)>
       where EvmConfig: ConfigureEvmEnv { ... }
}

A provider type that knows chain specific information required to configure an Env.

This type is mainly used to provide required data to configure the EVM environment.

Required Methods

fn fill_env_at<EvmConfig>( &self, cfg: &mut CfgEnv, block_env: &mut BlockEnv, at: BlockHashOrNumber, evm_config: EvmConfig ) -> ProviderResult<()>
where EvmConfig: ConfigureEvmEnv,

Fills the CfgEnv and BlockEnv fields with values specific to the given BlockHashOrNumber.

fn fill_env_with_header<EvmConfig>( &self, cfg: &mut CfgEnv, block_env: &mut BlockEnv, header: &Header, evm_config: EvmConfig ) -> ProviderResult<()>
where EvmConfig: ConfigureEvmEnv,

Fills the CfgEnv and BlockEnv fields with values specific to the given Header.

fn fill_block_env_at( &self, block_env: &mut BlockEnv, at: BlockHashOrNumber ) -> ProviderResult<()>

Fills the BlockEnv fields with values specific to the given BlockHashOrNumber.

fn fill_block_env_with_header( &self, block_env: &mut BlockEnv, header: &Header ) -> ProviderResult<()>

Fills the BlockEnv fields with values specific to the given Header.

fn fill_cfg_env_at<EvmConfig>( &self, cfg: &mut CfgEnv, at: BlockHashOrNumber, evm_config: EvmConfig ) -> ProviderResult<()>
where EvmConfig: ConfigureEvmEnv,

Fills the CfgEnv fields with values specific to the given BlockHashOrNumber.

fn fill_cfg_env_with_header<EvmConfig>( &self, cfg: &mut CfgEnv, header: &Header, evm_config: EvmConfig ) -> ProviderResult<()>
where EvmConfig: ConfigureEvmEnv,

Fills the CfgEnv fields with values specific to the given Header.

Provided Methods

fn env_with_header<EvmConfig>( &self, header: &Header, evm_config: EvmConfig ) -> ProviderResult<(CfgEnv, BlockEnv)>
where EvmConfig: ConfigureEvmEnv,

Fills the default CfgEnv and BlockEnv fields with values specific to the given Header.

Object Safety

This trait is not object safe.